Fix duplicate logging
Celery creates a StreamHandler on the root logger which results in
duplicate log messages because messages are propagated by each logger up
to ancestor loggers.

This commit disables the creation of all celery loggers by handling the
setup_logging signal. No loggers are created.

Another option was to set `propagate` to `False` on the root hope logger,
but that would only solve the problem for hope. If at a later date we'd
show logging messages from other libraries, the messages would be
duplicated again.